caliguland/caliguland-auth/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pyjwt==2.8.0
bcrypt==4.1.1
httpx==0.27.2
dstack-sdk==0.5.3